pyproject.toml
# Needed for UV Project management.
# Porn Fetch is NOT published in PyPi!

[project]
requires-python = ">=3.10,<3.14"
# 3.10 is needed due to truststore.
# You can manually patch it out from main.py and use older Python
# However, some libraries break due to type hinting!
# 3.14+ is not supported, due to Qt limitations.
# This is ONLY for the GUI, CLI is different

name = "Porn_Fetch"
version = "3.8"
authors = [
    {name = "Johannes Habel", email = "EchterAlsFake@proton.me"}
]

description = "a downloader for pornographic content"
readme = {file = "README.md", content-type = "text/markdown"}
license = "GPL-3.0-or-later"
license-files = ["LICENSE"]

dependencies = [
    "anyio",
    "beautifulsoup4",
    "brotli",
    "bs4",
    "certifi",
    "cffi",
    "charset-normalizer",
    "colorama",
    "fake-useragent",
    "h11",
    "h2",
    "hpack",
    "httpcore",
    "httpx",
    "hyperframe",
    "idna",
    "json5",
    "lxml",
    "m3u8",
    "markdown",
    "mutagen",
    "pycparser",
    "sniffio",
    "socksio",
    "soupsieve",
    "typing_extensions",
    "urllib3",
    "beeg_api",
    "eaf_base_api",
    "eporner_api",
    "hqporner_api",
    "missav_api",
    "phub",
    "porntrex_api",
    "spankbang_api",
    "xfreehd_api",
    "xhamster_api",
    "xnxx_api",
    "xvideos_api",
    "youporn_api",
]

# Note: This dependency list includes ALL dependencies of this project, and the dependencies of the dependencies.
# Usually you would only say: httpx, but httpx also depends on h11, httpcore and so on and I listed them all here.

[project.optional-dependencies]
gui = ["PySide6", "truststore", "pywin32; sys_platform == 'win32'", "pip", "cryptography"]
# pip is needed, because pyside6-deploy invokes pip when building the project with pyside6-deploy

cli = ["rich", "hue_shift", "pyinstaller"]
dev = ["rich", "hue_shift", "PySide6", "truststore", "pywin32; sys_platform == 'win32'", "pip", "cryptography"]
